Source Code/configurator/requirements.txt
pyserial>=3.5
pyinstaller>=6.0
Pillow>=9.0
esptool>=4.8
